requirements.txt
# Core AI and NLP dependencies
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Windows GUI automation
pyautogui>=0.9.54
pywinauto>=0.6.8
pygetwindow>=0.0.9
pillow>=10.0.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.1

# Memory and persistence
chromadb>=0.4.0
sqlite-utils>=3.35

# System operations
psutil>=5.9.0
pywin32>=306

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Logging and monitoring
loguru>=0.7.0

# Task scheduling
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
